+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   object, without a builder.
 - `woab::NamespacedSignalRouter` for routing signals from the same builder to
   different actors.
+- `BuilderConnector::with_object`.
 
 ### Changed
 - [**BREAKING**] Updated Actix to 0.11 and Tokio to 1.14. Consequences: